+++ rtic-monotonics/src/nrf/timer.rs
@@ -95,7 +95,7 @@ macro_rules! __internal_create_nrf_timer_interrupt {
 macro_rules! __internal_create_nrf_timer_struct {
     ($name:ident, $mono_backend:ident, $timer:ident, $tick_rate_hz:expr) => {
         /// A `Monotonic` based on the nRF Timer peripheral.
-        struct $name;
+        pub struct $name;
 
         impl $name {
             /// Starts the `Monotonic`.